Rename lhmn_ tables to lhma_ to avoid IBP stalls
When an LHM worker fails, `cleanup_current_run` must be called
to remove the triggers and "new" tables (which start with lhmn_).
The previous behavior was to drop the table
immediately. However, if this is an active table, the InnoDB buffer pool
can be full of pages related to this "lhmn_" table. When it is dropped,
this forces IBP to clear to those pages and can cause MySQL to become
unresponsive.

By instead renaming this table with the archive prefix (lhma_)
when can let the buffer unload relevant pages overtime, and then
later, safely, drop the archive tables as part of regular scheduled
maintenance.