Implement SqlRetry
This is a re-implementation of the RetryHelper that was introduced
in https://github.com/Shopify/lhm/pull/39.

By being a class instead of a module, it has a much more clear interface
and requires less setup. It features a true integration test for lock conditions
which were replaced in https://github.com/Shopify/lhm/pull/20 with mocks.

It's configuration is currently identical to the configuration of RetryHelper.