Ensure all table names originate from Lhm::TableName
As you can see, the logic for describing table names is all over the place.
Trying to change the interfaces for these classes is beyond the scope of the
PR, as these interfaces are sadly coupled to the test suite.

However, we can at least have these classes delegate the logic for these table
names to something is well tested and centralized. This eliminates the tests
currently running on migration_spec.